Token check to helper function
